requirements.txt
requests
pandas
python-dotenv
boto3
json